+++ packages/web3-types/CHANGELOG.md
@@ -215,3 +215,7 @@ Documentation:
 -   update the type for `baseFeePerGas` at `web3.eth.getFeeHistory` to be a number. (#7291)
 
 ## [Unreleased]
+
+### Added
+
+-   `FilterParams` type added (#7353)